pyproject.toml
[project]
name = "onshape-mcp"
version = "0.1.0"
description = "MCP server wrapping the Onshape REST API"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
onshape-mcp = "onshape_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/onshape_mcp"]
